+++ packages/desktop/views/onboarding/views/create-from-mnemonic/CreateFromMnemonicRouterView.svelte
@@ -4,7 +4,7 @@
     import { EncryptMnemonicView } from '../shared'
     import { CreateFromMnemonicRoute } from './create-from-mnemonic-route.enum'
     import { createFromMnemonicRoute, createFromMnemonicRouter } from './create-from-mnemonic-router'
-    import { VerifyMnemonicStatusView, VerifyMnemonicView, ViewMnemonicView } from './views'
+    import { VerifyMnemonicSuccessView, VerifyMnemonicView, ViewMnemonicView } from './views'
 
     $: if (features.analytics.onboardingRoute.enabled && $createFromMnemonicRoute) {
         Platform.trackEvent('create-from-mnemonic-route', { route: $createFromMnemonicRoute })
@@ -15,8 +15,8 @@
     <ViewMnemonicView />
 {:else if $createFromMnemonicRoute === CreateFromMnemonicRoute.VerifyMnemonic}
     <VerifyMnemonicView />
-{:else if $createFromMnemonicRoute === CreateFromMnemonicRoute.VerifyMnemonicStatus}
-    <VerifyMnemonicStatusView />
+{:else if $createFromMnemonicRoute === CreateFromMnemonicRoute.VerifyMnemonicSuccess}
+    <VerifyMnemonicSuccessView />
 {:else if $createFromMnemonicRoute === CreateFromMnemonicRoute.EncryptMnemonic}
     <EncryptMnemonicView router={$createFromMnemonicRouter} />
 {/if}
