+++ packages/desktop/views/onboarding/views/create-from-mnemonic/create-from-mnemonic-router.ts
@@ -19,9 +19,9 @@ export class CreateFromMnemonicRouter extends Subrouter<CreateFromMnemonicRoute>
                 nextRoute = CreateFromMnemonicRoute.VerifyMnemonic
                 break
             case CreateFromMnemonicRoute.VerifyMnemonic:
-                nextRoute = CreateFromMnemonicRoute.VerifyMnemonicStatus
+                nextRoute = CreateFromMnemonicRoute.VerifyMnemonicSuccess
                 break
-            case CreateFromMnemonicRoute.VerifyMnemonicStatus:
+            case CreateFromMnemonicRoute.VerifyMnemonicSuccess:
                 nextRoute = CreateFromMnemonicRoute.EncryptMnemonic
                 break
             case CreateFromMnemonicRoute.EncryptMnemonic:
