+++ packages/desktop/views/onboarding/views/create-from-mnemonic/views/VerifyMnemonicSuccessView.svelte
