+++ packages/desktop/views/onboarding/views/create-from-mnemonic/views/index.js
@@ -1,3 +1,3 @@
-export { default as VerifyMnemonicStatusView } from './VerifyMnemonicStatusView.svelte'
+export { default as VerifyMnemonicSuccessView } from './VerifyMnemonicSuccessView.svelte'
 export { default as VerifyMnemonicView } from './VerifyMnemonicView.svelte'
 export { default as ViewMnemonicView } from './ViewMnemonicView.svelte'
